+++ include/slang/ast/types/AllTypes.h
@@ -164,10 +164,10 @@ public:
     const Type& elementType;
     ConstantRange range;
     uint64_t selectableWidth;
-    uint32_t bitstreamWidth;
+    uint64_t bitstreamWidth;
 
     FixedSizeUnpackedArrayType(const Type& elementType, ConstantRange range,
-                               uint64_t selectableWidth, uint32_t bitstreamWidth);
+                               uint64_t selectableWidth, uint64_t bitstreamWidth);
 
     static const Type& fromDims(const Scope& scope, const Type& elementType,
                                 std::span<const ConstantRange> dimensions,
@@ -254,7 +254,7 @@ class SLANG_EXPORT UnpackedStructType : public Type, public Scope {
 public:
     std::span<const FieldSymbol* const> fields;
     uint64_t selectableWidth = 0;
-    uint32_t bitstreamWidth = 0;
+    uint64_t bitstreamWidth = 0;
     int systemId;
 
     UnpackedStructType(Compilation& compilation, SourceLocation loc, const ASTContext& context);
@@ -289,7 +289,7 @@ class SLANG_EXPORT UnpackedUnionType : public Type, public Scope {
 public:
     std::span<const FieldSymbol* const> fields;
     uint64_t selectableWidth = 0;
-    uint32_t bitstreamWidth = 0;
+    uint64_t bitstreamWidth = 0;
     int systemId;
     bool isTagged;
